\name{beckham.eff}
\alias{beckham.eff}
\docType{data}
\title{
amplification efficiencies for beckham.data
}
\description{
amplification efficiencies for beckham.data
}
\usage{data(beckham.eff)}
\format{
  A data frame with 11 observations on the following 2 variables.
  \describe{
    \item{\code{gene}}{a factor with levels \code{atf} \code{dnaja4} \code{dnajb1} \code{dnajb4} \code{egr} \code{fos} \code{gadd} \code{gapdh} \code{hsp110} \code{hspb} \code{r18s}}
    \item{\code{eficiency}}{a numeric vector}
  }
}
\source{
Beckham et al. Microarray analysis of cellular thermotolerance. 
Lasers Surg Med. 2010 Dec;42(10):752-65. doi: 10.1002/lsm.20983
}
\examples{
data(beckham.eff)
}
